requirements.txt
fastapi 
uvicorn
langchain-core
langchain-openai
python-dotenv
langgraph
langchain
langchain-community
pdflatex
redis
weasyprint